pyproject.toml
[tool.poetry]
name = "mi-proyecto-fastapi"
version = "0.1.0"
description = "API con FastAPI, PyMongo y pruebas con pytest"
authors = ["Tu Nombre <tu@email.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pymongo = "^4.6.0"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"